Add shared API layer
